// backend/src/shared/enums/errorCode.enum.ts
const enum ErrorCode {
  AUTH_EMAIL_ALREADY_EXIST = 'AUTH_EMAIL_ALREADY_EXIST',
  AUTH_INVALID_TOKEN = 'AUTH_INVALID_TOKEN',
  AUTH_USER_NOT_LOGIN = 'AUTH_USER_NOT_LOGIN',
  AUTH_USER_NOT_FOUND = 'AUTH_USER_NOT_FOUND',
  AUTH_NOT_FOUND = 'AUTH_NOT_FOUND',
  AUTH_TOO_MANY_REQUEST = 'AUTH_TOO_MANY_REQUEST',
  AUTH_UNAUTHORIZED_ACCESS = 'AUTH_UNAUTHORIZED_ACCESS',
  AUTH_TOKEN_NOT_FOUND = 'AUTH_TOKEN_NOT_FOUND',
  AUTH_INVALID_EMAIL = 'AUTH_INVALID_EMAIL',
  AUTH_INVALID_CREDENTIALS = 'AUTH_INVALID_CREDENTIALS',
  AUTH_USER_NOT_ACTIVE = 'AUTH_USER_NOT_ACTIVE',
  AUTH_USER_ALREADY_ACTIVE = 'AUTH_USER_ALREADY_ACTIVE',
  AUTH_USER_NOT_VERIFIED = 'AUTH_USER_NOT_VERIFIED',
  AUTH_USER_ALREADY_VERIFIED = 'AUTH_USER_ALREADY_VERIFIED',
  AUTH_PASSWORD_MISMATCH = 'AUTH_PASSWORD_MISMATCH',
  AUTH_INVALID_CURRENT_PASSWORD = 'AUTH_INVALID_CURRENT_PASSWORD',

  // Access control Errors

  ACCESS_FORBIDDEN = 'ACCESS_FORBIDDEN',
  ACCESS_UNAUTHORIZED = 'ACCESS_UNAUTHORIZED',
  BAD_REQUEST = 'BAD_REQUEST',

  // Database Errors
  DB_CONNECTION_ERROR = 'DB_CONNECTION_ERROR',
  DB_QUERY_ERROR = 'DB_QUERY_ERROR',
  DB_RECORD_NOT_FOUND = 'DB_RECORD_NOT_FOUND',
  DB_RECORD_ALREADY_EXISTS = 'DB_RECORD_ALREADY_EXISTS',
  DB_VALIDATION_ERROR = 'DB_VALIDATION_ERROR',

  // Validation Error
  VALIDATION_ERROR = 'VALIDATION_ERROR',
  RESOURCE_NOT_FOUND = 'RESOURCE_NOT_FOUND',
  RESOURCE_ALREADY_EXISTS = 'RESOURCE_ALREADY_EXISTS',

  // General System Errors
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  UNEXPECTED_ERROR = 'UNEXPECTED_ERROR',
  VERIFICATION_ERROR = 'VERIFICATION_ERROR',
  NOT_IMPLEMENTED = 'NOT_IMPLEMENTED',
}

export { ErrorCode };